+++ src/commands/ext-dev-emulators-start.ts
@@ -1,10 +1,6 @@
-import { checkMinRequiredVersion } from "../checkMinRequiredVersion";
 import { Command } from "../command";
-import * as controller from "../emulator/controller";
 import * as commandUtils from "../emulator/commandUtils";
-import * as optionsHelper from "../extensions/emulator/optionsHelper";
-import * as utils from "../utils";
-import { FirebaseError } from "../error";
+import { logger } from "../logger";
 
 module.exports = new Command("ext:dev:emulators:start")
   .description("start the local Firebase extension emulator")
@@ -14,28 +10,14 @@ module.exports = new Command("ext:dev:emulators:start")
   .option(commandUtils.FLAG_TEST_PARAMS, commandUtils.DESC_TEST_PARAMS)
   .option(commandUtils.FLAG_IMPORT, commandUtils.DESC_IMPORT)
   .option(commandUtils.FLAG_EXPORT_ON_EXIT, commandUtils.DESC_EXPORT_ON_EXIT)
-  .before(checkMinRequiredVersion, "extDevMinVersion")
-  .action(async (options: any) => {
-    const killSignalPromise = commandUtils.shutdownWhenKilled(options);
-    const emulatorOptions = await optionsHelper.buildOptions(options);
-
-    let deprecationNotices;
-    try {
-      commandUtils.beforeEmulatorCommand(emulatorOptions);
-      ({ deprecationNotices } = await controller.startAll(emulatorOptions));
-    } catch (e: any) {
-      await controller.cleanShutdown();
-      if (!(e instanceof FirebaseError)) {
-        throw new FirebaseError("Error in ext:dev:emulator:start", e);
-      }
-      throw e;
-    }
-
-    utils.logSuccess("All emulators ready, it is now safe to connect.");
-    for (const notice of deprecationNotices) {
-      utils.logLabeledWarning("emulators", notice, "warn");
-    }
-
-    // Hang until explicitly killed
-    await killSignalPromise;
+  .action((options: any) => {
+    const localInstallCommand = `firebase ext:install ${options.cwd}`;
+    const emulatorsStartCommand = "firebase emulators:start";
+    logger.error(
+      "ext:dev:emulators:start is no longer supported. " +
+        "Instead, navigate to a Firebase project directory and add this extension to the extensions manifest by running:\n" +
+        localInstallCommand +
+        "\nThen, you can emulate this extension as part of that project by running:\n" +
+        emulatorsStartCommand
+    );
   });