+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Removes support for running the emulators with Java versions prior to 11.
 - Removes `params` flag from ext:install, ext:update, ext:configure commands as they are replaced by the Extensions Manifest. See https://firebase.google.com/docs/extensions/manifest for more details.
+- Removes `ext:dev:emulators:start` and `ext:dev:emulators:exec` preview commands.